Fixes regex that forces a git ssh url to have .git
Also makes the username for the ssh url dynamic.

The git ssh url regex was not allowing URLs that do not end in .git to
be used. This commit allows a URL to be structures either as
git@github.com:example/example.git or git@github.com:example/example.

It also makes the username portion of the URL to be dynamic, which will
allow this feature to be used for many self-hosted git repo management
platforms like gitea and gitlab, where the username can be changed.