conditionals: removes Conditional and Expr models and updates equals logic
This commit removes the model after a brief redesign of what resolves
the expressions. Since the metadata-api will attempt to resolve the left
and right expressions, the arguments passed to fn.equals and
fn.notEquals should be simple Go types, so we use reflection to test if
the types are comparable and then do a simple equality operation on
them.